Split App into src
