+++ src/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+const styles = StyleSheet.create( {
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+} );
+
+export default function App(): JSX.Element {
+    return (
+        <View style={styles.container}>
+            <Text>Open up App.tsx to start working on your app!</Text>
+        </View>
+    );
+}
